examples/ai-and-agents/rag_knowledge_api/requirements.txt
azure-functions
azure-functions-knowledge-python
azure-functions-logging-python
azure-functions-openapi-python
azure-functions-validation-python
pydantic>=2.0
